feat: implement content filtering and file processing features
- Added content filtering functionality in `content-filtering.ts` with options for enabling filtering, processing filter types, and displaying statistics.
- Introduced file processing capabilities in `file-processing.ts`, allowing users to process various file types with detailed logging and error handling.
- Created input processing utilities in `input-processing.ts` to assemble queries from standard input and file content.
- Developed provider processing logic in `provider-processing.ts` to handle streaming AI responses with filtering options.
- Implemented response handling in `response-handling.ts` to manage the streaming of responses and save interactions to the database.
- Enhanced session management in `session-management.ts` to validate profiles and handle session states effectively.
- Added template processing functionality in `template-processing.ts` to apply act templates to queries.
- Updated help text in `help.ts` to include examples for new features like content filtering and batch processing.
- Defined command options in `options` directory for various functionalities, including file processing, filtering, and session management.
- Created a centralized index for command options to streamline option management.
