+++ src/cli/commands/ask/index.ts
@@ -1,0 +1,2 @@
+export { askCommandOptions } from './options/index.js';
+export { askCommandHelpText } from './help.js';
