+++ src/cli/commands/ask/options/files.ts
@@ -1,0 +1,7 @@
+import type { CommandOption } from '../types.js';
+import chalk from 'chalk';
+
+export const filesOption: CommandOption = {
+  flags: '-F, --files <paths...>',
+  description: chalk.white('Process content from multiple files'),
+};
