+++ src/cli/commands/ask/options/new-session.ts
@@ -1,0 +1,7 @@
+import type { CommandOption } from '../types.js';
+import chalk from 'chalk';
+
+export const newSessionOption: CommandOption = {
+  flags: '-n, --new-session',
+  description: chalk.white('Start a new chat session'),
+};
