+++ src/cli/commands/ask/options/verbose.ts
@@ -1,0 +1,7 @@
+import type { CommandOption } from '../types.js';
+import chalk from 'chalk';
+
+export const verboseOption: CommandOption = {
+  flags: '--verbose',
+  description: chalk.white('Show detailed status and debug information'),
+};
