+++ src/cli/commands/ask/types.ts
@@ -1,0 +1,5 @@
+export interface CommandOption {
+  flags: string;
+  description: string;
+  defaultValue?: string | boolean | string[];
+}
